+++ CHANGELOG.md
@@ -4,10 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/stefanzweifel/laravel-stats/compare/v2.2.0...HEAD)
+## [Unreleased](https://github.com/stefanzweifel/laravel-stats/compare/v2.3.0...HEAD)
+
+### Added
+- Add Database Factory Classifier [#183](https://github.com/stefanzweifel/laravel-stats/pull/183)
+- Add Support for Laravel 8 [#180](https://github.com/stefanzweifel/laravel-stats/pull/180)
+
+
+## [v2.3.0](https://github.com/stefanzweifel/laravel-stats/compare/v2.2.0...v2.3.0) - 2020-09-01
 
 ### Added
 - Add Observer Classifier [#177](https://github.com/stefanzweifel/laravel-stats/pull/177), [#128](https://github.com/stefanzweifel/laravel-stats/issues/128)
+- Add `--share` option [#178](https://github.com/stefanzweifel/laravel-stats/pull/178)
+
 
 ## [v2.2.0](https://github.com/stefanzweifel/laravel-stats/compare/v2.1.1...v2.2.0) - 2020-04-04